refactor: extract shared MANIFEST to scripts/example-manifest.js
The MANIFEST array was duplicated verbatim in build-examples.js and
the now-deleted add-example-chrome.js. Moving it to a dedicated module
eliminates the risk of the two scripts silently diverging when a crumb,
desc, or lectures value is updated in one but not the other.

Fixes #319
